chore: move cache due to conflict
