+++ backend/.env.sample
@@ -1,1 +1,3 @@
 NODE_ENV=development
+JWT_SECRET=your_secure_jwt_secret_here
+JWT_EXPIRES_IN=1h
