+++ backend/src/app.ts
@@ -8,7 +8,7 @@ import api from './api';
 import MessageResponse from './interfaces/MessageResponse';
 
 require('dotenv').config();
-import expressCache from "cache-express";
+
 
 const app = express();
 
@@ -23,22 +23,7 @@ app.get<{}, MessageResponse>('/', (req, res) => {
   });
 });
 
-app.use('/api/v1', expressCache({
-    timeOut: 60000, // Cache for 1 minute
-    onTimeout: (key: string) => {
-        const now = new Date();
-        const colors = {
-            reset: '\x1b[0m',
-            cyan: '\x1b[36m',
-            yellow: '\x1b[33m',
-            magenta: '\x1b[35m'
-        };
-        console.log(
-            `${colors.cyan}[CACHE]${colors.reset} ${now.toISOString()} - ` +
-            `${colors.yellow}Cache entry expired${colors.reset} for key: ${colors.magenta}${key}${colors.reset}`
-        );
-    },
-}), api);
+app.use('/api/v1', api);
 
 app.use(middlewares.notFound);
 app.use(middlewares.errorHandler);
